refactor(markers): namespace marker synthesizers under RbsInfer::Markers (#26)
Half-compact form preserves blame; analyzer call-sites and specs use
the fully qualified RbsInfer::Markers::* path. Require paths in
lib/rbs_infer.rb updated to markers/.
